Upload
WB stands for WithBag and the WOB for WithOutBag. The little boxes over
the inventory boxes are for the names. The first inventory part is for
pockets, which you spawn with. (if no clothes no inventory). The middle
one (only seen in WB) is the bag inventory. The buttom boxes are the
in-hands inventory. The items in in-hands inventory can be selected
(like in Minecraft) with numbers / scrolling with the mouse.